requirements.txt
dj-database-url==0.4.2
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
gunicorn==19.7.1
Markdown==2.6.9
pytz==2017.2
whitenoise==3.3.0
urllib3==1.22
python-decouple==3.0
dj-static==0.0.6
psycopg2
